More updates

+++ src/Sedan.java
@@ -2,7 +2,7 @@ public class Sedan extends Vehicle{
     private int range;
     private String engineType;
 
-    public Sedan(String brand, int year, float price, int range, String engineType){
+    public Sedan(String brand, int year, double price, int range, String engineType){
         super(brand, year, price);
         this.range = range;
         this.engineType = engineType;
