+++ src/Vehicle.java
@@ -1,9 +1,9 @@
 public class Vehicle {
     private String brand;
     private int year;
-    private float price;
+    private double price;
 
-    public Vehicle(String brand, int year, float price){
+    public Vehicle(String brand, int year, double price){
         this.brand = brand;
         this.year = year;
         this.price = price;
